pyproject.toml
[tool.poetry]
name = "simple_bucket_trigger"
version = "1.1.13"
description = ""
authors = ["george <georgelenton@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
google-cloud-storage = "^2.18.2"
google-cloud = "^0.34.0"
google-cloud-logging = "^3.11.3"
numpy = "^2.1.3"
tqdm = "^4.67.0"
google-cloud-bigquery = "^3.27.0"
loguru = "^0.7.3"
matplotlib = "^3.9.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pdbpp = "^0.10.3"
pre-commit = "^4.0.1"
pytest = "^8.3.4"
build = "^1.2.2.post1"
deptry = "^0.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.deptry.per_rule_ignores]
DEP002 = [
	"google-cloud",
	"matplotlib",
]
